mesure image & date
The image with the mesures will be saved if the user accepts the correction.
The name will be the same as the original image with .mesure before the extension. So images that end with mesures.png won't be taken as captures.

If the name of the file contains the date of the capture (like this: TEN_0000-M1.20220701.png) then the date will be saved in the landmarks file as 'capture_date': '2022-07-01'.