improved detection
Detection now doesn't use a custom neural network, instead it uses MediaPipe in two ways, depending on the point:
· It interpolates the points from MediaPipe Hands.
· It follows a line that passses throug two MediaPipe Hands' points until it finds a significant color change, i.e., the edge of the hand.
This turns out to be more accurate than the custom neural networks that we had (that performed poorly).
Is more robust, since the only ML part is MediaPipe Hands, that has been trained in a wide dataset.
It can be runned without TensorFlow or Keras, it simply needs MediaPipe. (We had a proble with loading a NN to Keras due to a problem with the h5py library).

The code on handmeasure is a bit more clear.

The images with the measurments are now saved as .jpg.
This makes the code cleaner (we only have to check for the extension) and saves a lot of disk space.

Now JSON files don't have the '_lm' disctintion on the name, they are just JSONs.