sticky edges
When pressing the Shift key the points will stick to the closest edge.

When perssing the key 'g' the hand will be saved. It has the same behaviour as space.

In the initial estimation of the keypoints now they are way less likey to end up in the wrong finger.